pyproject.toml
[project]
name = "semantic-mail"
version = "0.1.0"
description = "Local semantic search for Gmail using Ollama embeddings"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.0.13",
    "click>=8.2.1",
    "google-api-python-client>=2.173.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "ollama>=0.5.1",
    "openai>=1.88.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "tqdm>=4.67.1",
]

[project.scripts]
smail = "src.cli:cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
